refatora agentes para suporte a conhecimento vetorial; substitui chamadas de funções e ajusta instruções de interação
