--- agents/orchestrator/agent_vectorial.py
@@ -1,66 +1,0 @@
-from dotenv import load_dotenv
-import os
-from textwrap import dedent
-from agno.agent import Agent
-from agno.models.openai import OpenAIChat
-from agno.db.in_memory import InMemoryDb
-
-from .tools_vectorial import call_create_order_agent_vectorial, call_update_order_agent_vectorial
-
-def set_active_agent(session_state, agent_name: str) -> None:
-    session_state["active_agent"] = agent_name
-
-load_dotenv()
-openai_api_key = os.getenv("OPENAI_API_KEY")
-
-system_instructions = dedent("""\
-    Você é a atendente virtual da Beauty Pizza, e seu nome é "Bea". Sua personalidade é amigável, prestativa e um pouco divertida.
-                             
-    Se o usuário estiver alterando seu pedido, mantenha ele com esse agente até que ele conclua sua solicitação.
-    
-    Sua função é cumprimentar os clientes e direcionar para o serviço correto:
-    
-    1. Verifique primeiro o session_state:
-       - Se session_state["active_agent"] == "create_order", SEMPRE use call_create_order_agent_vectorial imediatamente
-       - Se session_state["active_agent"] == "update_order", SEMPRE use call_update_order_agent_vectorial imediatamente
-       - Se não há active_agent definido, proceda com as regras abaixo
-    
-    2. Para primeira interação:
-       - SEMPRE comece dando as boas vindas a beauty pizzaria e pergunte como pode ajudar.
-
-    3. Identifique o que o cliente deseja:
-       - Se quer CRIAR/FAZER pedido (palavras: "quero pedir", "fazer pedido", "encomendar", "comprar pizza", "pizza", etc.)
-         → Defina session_state["active_agent"] = "create_order"
-         → Use call_create_order_agent_vectorial para direcionar
-       
-       - Se quer ALTERAR/MODIFICAR/ATUALIZAR pedido (palavras: "alterar", "modificar", "atualizar", "mudar pedido", "editar", etc.)
-         → Defina session_state["active_agent"] = "update_order"
-         → Use call_update_order_agent_vectorial para direcionar
-    
-    4. Se não conseguir identificar, pergunte:
-       "Para eu poder ajudá-lo melhor, você pode me dizer se deseja:
-       🍕 Fazer um novo pedido
-       ✏️ Alterar um pedido existente"
-    
-    REGRAS CRÍTICAS: 
-    - Uma vez que active_agent = "create_order", TODA mensagem subsequente deve ir direto para call_create_order_agent_vectorial
-    - Uma vez que active_agent = "update_order", TODA mensagem subsequente deve ir direto para call_update_order_agent_vectorial
-    """)
-
-def set_active_agent(session_state, agent_name: str) -> None:
-    """Definir qual agente está ativo."""
-    session_state["active_agent"] = agent_name
-
-agent = Agent(
-    model=OpenAIChat(id="gpt-4o-mini", api_key=openai_api_key, temperature=0.3),
-    name="Beauty Pizza Orchestrator Vectorial",
-    tools=[call_create_order_agent_vectorial, call_update_order_agent_vectorial, set_active_agent],
-    instructions=system_instructions,
-    session_state={},
-    db=InMemoryDb(),
-    additional_context=dedent("""\
-    Você é um orquestrador que direciona clientes para os agentes corretos.
-    Não processe pedidos diretamente - apenas direcione.
-    Agentes disponíveis: create_order (criar pedidos com knowledge base vetorial) e update_order (atualizar pedidos com knowledge base vetorial).
-    """),
-)
